requirements.server.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.1
python-dotenv==1.0.0
pydantic
aiohttp==3.9.1
httpx==0.25.1
aiofiles
numpy
colorama
safetensors
transformers<4.54.0
torch==2.8.0
huggingface_hub[hf_transfer]
mistral-common>=1.7.0
